requirements.txt
torch
torchvision
pillow
ultralytics
fastapi
uvicorn
python-multipart
